backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
python-dotenv
alembic
passlib[bcrypt]
python-jose[cryptography]
httpx